backend/requirements.txt
pyodbc
sqlalchemy
fastapi
uvicorn
pandas
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openpyxl
psycopg2-binary
python-dotenv